# Searching on Binned Edges