… test/tiles/2/000/519/120.gph