@kevinkreiser 2016-03-23
closes #52